Update README.md
Added a server template in the documentation for Debian 7 / Ubuntu 14.04 / CentOS 6.5 / CentOS 7 and Amazon Linux. This allows users to:
 - Have working instructions to install the application from scratch 
 - Create a temporarily server with a working version of tutanota to play around
 - Deploy a production server with tutanota

I think this is a great add-on for the documentation. Novice users will be able to run the application with no problem and advanced users has templates that they just need to customise.

Feedback is appreciated!