[dev] Fix importing files during bootstrap
This commit fixes loading the debug build in Safari and occasional
failure to load in other browsers.

node server didn't work with Safari, I noticed that it tries to load js
files without .js extension. I compared debugging with Chrome and
noticed that we do some replacement in system-resolve.js. This was not
running in Safari and in fact system-resolve was not loaded at all in
Safari. I checked where it's imported and there was an error in make.js
where it was not waiting for the import of system-resolve to import
src/app.js. This was a timing issue and it explains why we had such a
problem in other browsers but only occasionally: sometimes it was fast
enough to patch systemjs when it was from disk/code cache.