Fix folder more button animation glitch.
The button was jumping around. It was introduced in 9d7a8df where the
createMailBoxExpanderButton() call was moved inside the view() function
to create different buttons and be independent from the currently
selected folder. Because of that mithril was thinking that this is a
different component and was replacing it so we had two (or more)
buttons at the same time in a single row (their removal is delayed
until the animation end).

The old-style button was replaced with new-style one which doesn't
cause this issue because the "tag" (class) stays the same between
redraws and mithril knows that this is the same component.