[android] Fix generating RSA key on newer Android 9. Fix #1629
The last place we explicitly specified BouncyCastle was RSA key
generation. It started failing after we started targeting newer
Android in 3cf81f0. We do not specify provider anymore and it should be
BC on older versions and Conscrypt/AndroidOpenSSL on newer ones.
Specifying the provider is not recommended in any case and we got it
working for other operations so it should not be an issue.

See https://android-developers.googleblog.com/2018/03/cryptography-changes-in-android-p.html