Fix calendar events on being visible on 1st day of month, fix #1730
The problem could be reproduced in timezones west to UTC. All-day
events have start time as UTC 00:00 which falls on the previous day in
time zones which are negative to UTC (e.g. MST/UTC-7). We load events
with overlaps and later filter out only those which belong to the
month and because all-day events have been falling on the previous
month (if you don't convert time to local) they've been filtered out.

This commit converts start time of the event correctly.