Fix display problem caused by negative margin, fix #1917
Changes in Blink caused our text fields with buttons to look incorrectly
after focusing one of the buttons. This happened because we used
negative margin to align button to the right edge (to the text field
and other component).

Instead we now align button contents to the right in text fields.
There are still couple of places where negative margin is used but
there are no issues with it so far.

This commit changes old-style DropdownSelector to use new-style buttons
internally to avoid implementing endAligned for the old-style Button.

This commit also changes buttons to the new ones whenever possible.