Fix receiving u2f error after login was accepted, fix #1792
This happens because we receive generic U2fError instead of timeout
error (seems like u2f has its own timeout). Now we will only show the
error dialog if the waiting dialog is still visible. We would like to
cancel u2f request but it doesn't seem to be possible.