Fix login in IE. Fix transpiling of embedded libs, fix #2035
There are multiple errors which caused this.

One is debug build which was not transpiling anything in the worker
libs folder. Exception was made for the /libs folder in the root but
it affected the other /libs folder as well.

Another is related to the import in murmurhash. It was done in the form
of const ... = require(...) and "const" was left as is after the
compilation.