[android] Use single user when opening SSE connection, #2362
Previously we've been passing all user ids in the hope that one of them
will match. Now we would like to change this and pass only single user
id and if it fails delete it and try the next one.

This commit also removes some of the old migration mechanism from
SseStorage.