[desktop] enable context isolation for all BrowserWindows
The preload script required some access to the window object of the renderer
thread to implement browser functionality that's missing in a BrowserWindow.
This prevented use of the contextI isolation option.

This commit moves that functionality from the preload script to either the main
thread or the web app so we can enable context isolation.

* mouse wheel zoom was moved to the main thread, listening to the new
 zoom-changed event on WebContents (ApplicationWindow.js)
* url reveal on hover was split into a main part listening on update-target-url
 on WebContent (ApplicationWindow.js) and a new NativeWrapperCommand to fill in
 the url element
* the fix for MacOS navigation history when trying to navigate a text box with
 cmd+arrows moved to WindowFacade.js
* the SearchPageOverlay now fixes the non-detection of focus change itself
* unregistering the window on unload moved to ApplicationWindow.js, detecting
 a reload in the did-start-navigation handler
* the fix for the nonfunctional window focus was removed since it works in current
 electron versions

all windows now use the same ipc channel, the IPC component uses the Event sender
to route responses back to the correct window, removing the need to set up
the channel with the window id