Fix "batch does not have events for" error, fix #2648
In case of DELETE we have been removing all the batches between it and
previous DELETE/MOVE. We did not add the new DELETE because the old
one should be enough. However those operations inbetween can still be
"latest for entity", even if they are empty and since we don't add a
new DELETE to the queue we wouldn't operate latest entry in the map.

This commit sets previous batch as the latest one instead of the
removed batch