Fix expander panel in the user settings view, fix #3501
The problem was that the expander panel was expecting the `expanded` stream to be the same one everytime it was passed in. This turned out not to be the case, because the UserViewer would be recreated whenever the selected user was changed, but mithril would reuse the existing expander since the layout hasn't changed significantly.

Now we handle removing the expander children from the dom based on when it's transitions start and end, so we don't rely on the stream being the same.