Refactor date picker and deal with the consequences
We strived to not have to use streams for the start and end date of the CalendarEventEditDialog, but some issues came up, particularly how errors would be handled when parsing text. Fixing these cleanly required touching all the code that used date pickers, which is why this became so much bigger than one might expect. the date pickers do now behave in less surprising ways