[desktop] make resolvePushIdentifierSessionKey more robust
alarmnotifications may come with pushIdentifierSessionKeys
that are not accessible anymore due to lost deviceKeys

the alarmscheduler may try to decrypt them anyway, which
can lead to a situation where an alarmnotification
continually causes errors.

with this commit, the scheduler will try keys until it has
none left before resetting the SseInfo and redownloading
all alarms.