[desktop] try to find perfect match before fallback textId in SpellcheckLanguageDialog.js
SpellcheckLanguageDialog.js looked for the first language in the
LanguageViewModel that fullfills any of these:
* its langCode-locale has perfect match with dictionaries langCode-locale
* it only has a langCode that matches the dictionaries' langCode perfectly
* the language code with its locale removed matches the dictionaries' langCode

the language pt-br runs into the third case and gets returned for the pt-pt dict,
even though later in the languages list there would be a perfect match with pt-pt.

this commit adds another missing language text id (Portuguese) and changes
the textId selection so it tries to find perfect matches before considering
any fallbacks.

fix #3659