Fix initializing cache with batch ids for the first time
For each group we record the last batch we processed so that we can
download events since that batch and update the cache. Unfortunately
we would only write those ids when some events are processed. In case
there are no events we would not download any batches on next login.

This fixes plugs the issue with this case by preemptively writing last
id if none could be loaded. Next time we will download batches starting
with that id.