[android] Fix hang on opening email notifications, fix #3911
Android would fail to resume activity because we tried to send message
to web thread way too early, before the page would load or before the
bridge would be established. This only surfaced with 7fa44f because
we calling methods on message port would result in an error, before it
would just try to evaluate js into nothing.

We changed the implementation of sendRequest() to wait for the app and
bridge to be initialized so that no external call can call the method
incorrectly anymore. This should prevent this error in the future but
it still seems to be pretty brittle because there's no proper null
checking and because of how unwieldy old promise library is.