[android] Fix crash caused by IOUtils update on old systems, fix #3962
Newer IOUtils requires Android 9 or at least some APIs from it and
Android 6 and 7 do not have it without some kind of polyfills.

It would manifest as constant crashes for background process and
crashes when downloading files.

We need to either move away from io-commons or introduce some sort of
de-sugaring/polyfill.