allow import of events with illegal end time
events conforming to https://tools.ietf.org/html/rfc5545 must have
an end time that is larger than the start time.

Some applications seem to export "reminder" like events with identical
start and end times.

we can safely provide some functionality anyway by assuming a legal
endTime if the event is otherwise okay.

illegal all-day events (ie events that only specify a start day,
but no time) will be converted to an all-day event with a duration of
one day.

illegal events that are not all-day will be converted to a duration of
1 second for the following reasons:
* duration is rendered as eg 2022-03-15 17:00 to 2022-03-15 17:00
* the duration has the minimal chance to overlap the end of day which
would cause the event to be rendered over 2 days
* we don't lose alarms by importing an event that would start in the
future as an all-day event that has already started

fix #3764