Fix AddSpamRuleDialog being stuck in invalid state, fix #4152
AddSpamRuleDialog would try to load some data for validation but would
never expect an error if it can't load it. In that case it would return
an "emptyString_msg" (which denotes "failed" validation) so the user
would see an empty dialog.

Now we catch the errors while loading the data and close the dialog if
it fails as it's the easiest way to trigger retry. Also failed
validation dialog would display loading message instead of an empty
dialog.