Fix logging in after upgrading from 3.96.3, fix websocket, fix #4150
The problem occurred because before offline login we didn't use the
cache for the login process so user's GroupInfo was likely not cached as
it is not downloaded that often after login.

Now there's a temporary workaround that just falls back to the sync login.

Also, after e3fc4135be9e8b855fb876ccb88476f39b2f04f6 websocket would not
connect in async login because it would think that there's already
previous login (because of the partial login), so it would try to
reconnect instead which didn't work because we started in the
terminated state.

Now it checks for full login and always uses connect().