Fix incomplete login error handling, fix #4165
After we added offline login functionality it is possible to run into
a situation where we are online but we are not fully logged in (e.g.
if we didn't get the offline event, we tried to log in but failed or
when we are still logging in).

Most places do not need to differentiate between this case and
connection errors so we introduced a predicate so that we can easily
reuse our offline handling. This allows few places to still explicitly
differentiate between these kinda of errors.