Fix handling for expired sessions, fix #4193
With offline login we've added handling for closed session after going
online. The code path is the same as for expired sessions. For offline
we've added code which re-saves the credentials after new session is
created. This shouldn't happen for ephemeral sessions (when they expire)
because we don't store them but this check was missing.