Fix MailView assuming that the list is always there, fix #4192
We've been for a long time plagued by the timing issues with MailView.
The reason for this is that it was always assumed that we have a
MailListView. To have a MailListView we need to get a mailListId which
we only get from updateUrl which is executed async. Moreover, if no id
was given (as is the case for opening mailbox for the first time) then
we need to load some data from the server to find our mailbox id. If
the connection is slow enough it could lead to a "dead zone" where we
couldn't do anything to a MailView.

Passing list id as an attribute would be already an improvement as we
wouldn't need to rely on updateUrl but if there's no id we would anyway
have to wait.

We made the list explicitly nullable to address this case.