Fix unintended EntityRestCache behavior
EntityRestCache would process both LET and ET CREATE updates as LET
updates. In practice the only effect of this would be an erroneous
check for range with empty list id. We wouldn't download the entity
because there's no such range so no unnecessary network requests were
made. Still this was unintended and it should have taken ET path.

Co-authored-by: nig <nig@tutao.de>