Replace InfoView with NotFoundPage
There are 3 usages of InfoView  (one in the admin client)

the app.ts and admin client usages were identical, then only difference in ContactFormView was that no back button was rendered. I collapsed all three to use the same implementation and made InfoView into a more specific NotFoundPage component. The main difference from before is that ContactFormView now shows a back button in the 404 page, which seems fine to have at the benefit of a simpler component

#2559