fix link handling in mail viewer
touch and click handlers were attached to the wrong element
outside of shadow dom. Even when listener is set to the wrapper inside
the shadow dom event.target turns into pumpkin (shadow dom parent)
after the event loop. We handle both double tap and links by cancelling
real click events and then re-dispatching them after timeout (when we
are sure that it isn't a double-click). We changed how we dispatch it
because we don't have access to the original shadowed anchor element
anymore. Instead we just create a disposable anchor tag and to dispatch
the click event.

fix #4372

Co-authored-by: ivk <ivk@tutao.de>