Fix SelectMailAddressForm showing invalid state for a short time
verifyMailAddress() was not checking whether the state is up-to-date
when switching off loading state. This is a consequence of a loose
state representation. Because of that users would see "invalid" state:
"busy" would be turned off before the actual result would show up for
the new email address.

Ideally this should be re-written as a proper state machine and
extracted into a ViewModel or similar.

fix #4403