[android/ios] explicitly disable u2f for mobile apps
after the changes to the asset loading, the android client started
attempting u2f challenges because the domain matches the registered
keys.

this commit disables u2f in BrowserWebAuthn when used in apps until u2f
is properly supported.

this also fixes the misleading message
"security key is not registered for this domain" that would be shown to
mobile users, it's just "accept login from another client" now.

close #4426