Have MailView use MailUtils for modifying the unread flag
We already have a function in MailUtils for marking an array of emails
as unread or read, so it makes more sense to use this function for
this.

This function also checks the status of the emails before sending a
request, thus we do not need to send unnecessary requests such as
requesting an already read email be marked as read.

Fixes #4455