Fix deactivating a whitelabel date not sending a date
On commit 9d5e2f6fb4bc14a5cc0f7b1a74a606b32cd56e16, the visual status
of whitelabel children was corrected, but the actual effects of the
button were then broken, as it was written with the inverted
functionality (thus with this commit, deactivating now activates and
vice versa).

This commit fixes this issue by now having it send a deactivation date
on deactivation.

Fixes #4462