MailEditor should show CC/BCC if initialized with either
We want to prevent that, when sending emails, emails are unexpectedly
sent to recipients that the sender is unaware of.

Therefore, when opening a mail editor with pre-set CC/BCC recipients
such as when clicking on a mailto link or opening a draft, we want the
CC/BCC fields to be expanded automatically to indicate to the user that
there are more recipients.

Fixes #3844