Fix unwanted autologin when cancelling dialogs on login screen
When credentials are stored and the user cancels:

- the dialog for signing up a new account
- the dialog for taking over an account due to inactivity
- the dialog for having lost access to an account

the desired behavior should be that the user is not logged back into the
previously saved account.

To ensure this the parameters for redirecting to the login page have been
changed to not log the user in automatically after cancelling the mentioned
dialogs.

fixes #2549
Co-authored-by: @nif-tutao