[maintenance] remove node-forge dependency
node-forge is a pure JS crypto library that was used during the build
process to sign the clients and at runtime to verify the signature
before downloading the updated client:
* check the update manifest on the server, if the version is higher
  than installed:
* get the sha512 digest and signature from the manifest and verify the
  signature (node-forge only requires the digest, not the data)
  if it matches:
* download the update and let electron-updater do its thing

with this change, we have to download the update first because node
crypto requires the data to verify the signature (digest is calculated
internally). In case the signature doesn't match, we delete the
downloaded file and stop polling for updates for this run.

forge was also used to get a private key for signing out of a
self-signed cert in development environments. from now on, this is done
by providing PEM encoded private and public keys.

close #3832