Change picture on final signup
The picture on the final signup should be a different one.
Added the new encoded picture and introduced variable to use it in the
desired places.
"Icons.ts" which includes the image is auto-generated and must not
necessarily be merged.

!! Must also merge the commit in tutao-next repo which introduces the
code for the picture to be available in the Icon.ts file via auto-gen !

fixes #3989