Fix recovering session when offline mode is enabled
With refactoring of SQL-related code we introduced a bug where offline
db state errors produced during session recovery (when the session is
no longer valid, and we try to create a new one) would not be ignored
as before because we wouldn't throw OfflineDbClosedError which is
introduced specifically for this case.

Now we throw the correct error for both Android and Desktop and the
expected offline errors will be ignored.

fix #4527