Fix issues with temporary login
The first issue is attempting to get last update time in
OfflineIndicatorViewModel. It was trying to wait for partial login but
unfortunately temporary login is also partial login.

We didn't want to just return null in this case as it might get
interpreted incorrectly and instead we returned a sum type with the
exact case and ignore uninitialized storage in this case.

The other issue was with LoginController and LoginListener not properly
resetting themselves on logout (e.g. after temporary login) and giving
incorrect results.

The third issue was with PostLoginActions initializing MailModel for
temporary login. If a real login into another account is attempted
afterwards it would lead to more errors.

To fully solve the problem we should rethink how temporary login affects
the rest of the app and perhaps skip most of the things that happen
during normal login.

fix #4559