fix search on Android stops working after system back button press
The system back button on Android leads from the detail search view back
to MailView, when it should go back to the search overview

SearchView now overrides handleBackButton() like ContactView and
CalendarView do.
This method handles the back button only in single-column layout, as
this behavior is not needed for multiple columns.

fix #4621