Remove entities when the owner group membership is lost
Add ownerGroup column to offline db.

We detect the relationship change by the update of the user. We track
current user in cache storage because it already was receiving it in
case of offline storage and storage already had appropriate lifecycle
for it.

We empty the tables instead of migrating them because we want to remove
them anyway due to corruption (see #4664)

fix #2874