[ios] make sure the xcode run schemes access the correct web app build
the ios app accesses its UI from a path in the bundles asset folder that
is configured in the Info.plist file. This path was overridden by the
"tutanota" scheme to use the debug web app build produced w "node make".
This was done with an Environment variable defined on the scheme, which
is set on each run through xcode but not on subsequent launches from the
iPad UI, causing the app to show a white screen if the dist build of the
web app is not present in the build folder, possibly next to the debug
build.

this commit changes the run schemes of the project so the tutanota-dist
scheme always uses build/dist as a path (as configured in Info.plist)
and the tutanota scheme uses a pre- and post build script to modify
Info.plist during the build.

(this cost us about 12 hours)

Co-authored-by: nok <nok@tutao.de>