Fix grammar of no news message
The word 'news' is an uncountable noun, and so it will always get a singular treatment in English grammar, not plural treatment.
Other options to fix the grammar would be:
- There are no news items
- There are no new messages
While the first alternative would perhaps be closest to the German source, I think the correction suggested in this change (there is no news) is the most common formulation and still quite close.