fix android notification shows only the time even for future events
Previously, Android notifications only showed the time of an event, even
if the notification was for an event in the future and not on the same
day. Now, the notification text is generated depending on whether the
notification is for the same day or not. On desktop this is already
done in a similar way, see AlarmScheduler.

fix #3519