fix suggestion dropdowns being positioned to random ancestor
the suggestion dropdown is an absolutely positioned element,
these get positioned relative to the next positioned ancestor.

The MailRecipientsTextField did not come with its own
positioned element, so positioning of the dropdown depends
on the environment it's displayed in.

this commit adds a position: relative wrapper around the dropdown

also removes overflow: hidden from Expander since it's cutting off the
dropdown and doesn't seem to do anything in any browser (the issue
mentioned in f9796135229c588b4c5b7e0d6c3a6019f354028c was visible in the
new mailviewer header expander, but it's currently not anymore even
without the overflow: hidden)

fix #4725