fix modal doesn't close when using browser back button
For the whole app only one main modal is created which is then either
displayed or removed. When for example the ViewSlider is created after
the initial modal was created, it is not possible to close the current
modals using the browser back button due to the order of the event
handler stack.

#4731