Implement subscription footnotes on website
The SubscriptionSelector shows a few notes on the subscription at the
bottom. These notes are not on the website because the translation keys
for those notes are not fetched by the website.

This commit changes the translation keys to start with "pricing." so
that the website can fetch them and display them accordingly.

fixes #21