fix mail address is selected when closing mail recipient suggestion
When no recipient from the suggestions is explicitly selected, none of
the suggestions should be put to the recipient field. But currently when
closing the recipient suggestion dropdown by for example clicking
somewhere else the onBlur method is called and the current highlighted
suggestion gets selected.

Still the input needs to be resolved, even if the dropdown is closed.
This is for example necessary for external mail addresses.

fix #4831