[desktop] dispose DB before waiting for other windows to close it
if there are multiple windows logging out, we may have to wait for a bit
before we can safely delete the offline database.

previously, we waited before disposing, which means that the reference
held by the window that called deleteDb would always be held the full 1s
before the dispose/delete call.

#4866