Fix attempt to download all attachments when offline
Currently, attachments are downloaded one after the other, regardless of
whether errors occur. If the previous attachment failed to download
because the client is offline, there is no point in downloading more
attachments and making the user wait 20 seconds for each attachment.

Once a download fails because the client is offline, all subsequent
attachments will not be downloaded. However, if it was possible to
download attachments before, they will still be processed.

Downloading in the browser (download all only) and natively is now
combined in the doDownload function in FileController.

fix #4937