Fix when onNewUrl() is called, fix #4998
With old view resolver updateUrl() would always be called before the
view() method. With new view resolver and BaseTopLevelView this has
changed and it would be called afterwards. When this got changed
CalendarView would update it's state too late and would once redraw
with the old state.

This commit reverts to the previous behavior so that we have a chance
to set the state depending on the URL before we re-render.