Focus password field after signup.
Previously when the login form was created, the promise that stores the
first created login form was already resolved before the new handler
that focuses the password field was attached.
Immediately afterwards, we created a new loginForm and re-resolved the
promise, which caused the focus to be set on the just removed login form

re-creating the deferred object in the oninit of the login form allows
to focus the correct DOM Element.

#4630

Co-authored-by: nig <nig@tutao.de>