ignore autofilled mail address if it's empty
we had a fix for ios that sets the mail address on the loginForm
after it has been auto filled. this caused us to override the loginWith
value of the route before it was set. on platforms that don't autofill,
the override happened with an empty mail address.

we now check if the mail address is different and if it's not empty
before overrriding the value of the field.

fix #4630

Co-authored-by: nig <nig@tutao.de>