fix: typed in mail address is not being replaced after signup
The autofill handling in LoginForm prevents us from just changing the
mail address forms stream value. We need to change the fields to an
empty string (which won't be an autofill suggestion) to replace it
then with the actual mail address using the stream. Same procedure is
done to empty the password field.

fix #4630