don't create a separate passthrough stream for ProtocolProxy responses
we've been calling .end() in case of errors, which means the other side
can assume that everything is fine with the response and proceed.

we now directly pass the response readable to the sendResponse callback
and let node/electron handle the lifecycle of the streams, which should
result in a xhr.onerror callback call in the rest client when something
goes wrong.

close #5096