Fix getting into invalid login state
Login state is tracked differently for worker and page sides of the app.
When `LoginFacade` (worker) thinks that the partial login is done
`LoginController` (page) can still fail, mostly due to
`MailModel#init()` being part of the login process.

We suspect that the issue happened due to page login failing. On the
next try we would try to log in again even though the login is already
done from the worker's perspective.

We had two main options: either aborting worker login when page login
(including `PostLoginActions`) fails or taking `MailModel#init()` out of
the login process. We went with the latter because most of the code was
already written with the assumption that `MailModel` is initialized
asynchronously. This has an added benefit of letting user into the app
faster. Later we can add some placeholders for the MailView to improve
the experience even further.

Since `MailModel#init()` is not part of the login anymore we had to
find a way to retry the loading if it fails. We've adapted `MailModel`
to make the retry work (which we would need to do in either case) and
added another `init()` call on full login as well on using the
mailModel. We've changed `MailModel` to not propagate the init errors
to the rest of the app but to make them wait. This works better for
most places because we don't need additional error handling.

fix #5094