Fix uncaught error because of invalid isDetailsDraft() check
`isDetailsDraft()` would compare against the `null` strictly while in
some cases the field would just not be set (probably from cbor). It
would later proceed to try and use it in OfflineStorage#deleteMailList()
which would throw an assertion error.

We fixed the check to match both.

Ideally we should also update the generated types to account for that.

fix #5108