make disposing MailViewerViewModel the second time a noOp
the conversation view disposes us twice if our mail is deleted because
it's getting disposed itself (from the list selecting a different
element) and because it disposes the mailViewerViewModel that got
updated.

we might improve on this in a smarter way in the future depending on how
we handle entity updates changing the list of entity listeners before
they have been fully handled.

fix #5189