Patch mithril to fix unwanted logging behavior
We've discovered that mithril is swallowing up all uncaught errors
and is just logging them. This might be the best solution in some cases
where the app would be completely unusable otherwise. We can't do much
about it at the moment.

Additionally, console object was bound too early for us to add these
errors in the logger. To address this we've patched mithril so that it
doesn't keep its own copy.

#5221