fix not being able to create contacts from mail view
we assumed that the contact was already on the server if it was passed
at all, but in this case, we pre-populate a contact without an id
and pass it to the editor.

Trying to update that causes a 404 which went ignored.

we now check for presence of a contact and its Id to determine the
isNewContact flag and use that to decide whether to update or create.

fix #5232