Rewrite calendar editor and calendar popup
make updating calendar events delete the old uidIndex correctly

when updating calendar events, we sometimes had events in the
delete call that do not have the hashedUid set, causing us to
be unable to delete the uid index entry. when re-creating the
event, that leads to a db.exists error.

make sure the event popup always has the current version of the event.

now that edit operations are possible from the popup, we either need
to close the popup after it calls the model factory to make sure it's
only called once, or make sure the model factory always uses the last
version of the event.

we opted for the first option here to make sure repeated changes
to the attendance are actually sent as a response.

make contact resolution failure more controlled for external users

previously, responding to an event from an external mailbox
would try to resolve a contact for the response mail, fail,
try to create a contact and fail again on an opaque
assertNotNull.

show partial editability banner when creating new event in shared
calendar

make it possible to add alarms to invites in private calendars

don't make saving/sending invites and cancellations depend on user
choice for own events

updateExistingEvent() should call sendNotifications/saveEvent even when
there are no update worthy changes or the user did not tick the
sendUpdates checkbox because invites/cancellations must be sent in
any case and sending updates has a separate check for sendUpdates

also make the sendUpdates button on the popup use the same logic
as the shortcut when clicked (ask confirmation)

Co-authored-by: nig <nig@tutao.de>