make ArrayUtils.partition a type assertion, add mapMapAsync
sometimes, `partition` would be useful to act as a type guard to save some
repeated checks, like partitioning on the presence of a field and then
operating on that field. usages where both resulting arrays are the same
type are not affected.

mapMapAsync maps a maps contents from one type to another while leaving the
keys alone.