Fix deleting events with external participants from calendar event popup
CalendarEventWhoModel would set the external password for recipients
even when it wouldn't have one yet (and in case of event popup it
wouldn't have enough time to resolve the contact and get the password
from it). SendMailModel wouldn't set the password from the contact if
it already had one set so we would end up with an empty password
because of the timing.

Making sure we don't set the password unless we readlly have it fixes
the issue.

#5437