do not cache CalendarEventUidIndex entries
some calendaring apps send the same event
multiple times in the same mail, which
causes us to handle the update twice.

this is not a problem as long as the
entry is up-to-date, but since entity
events take a while to arrive, we
need to make sure the cache is not storing
these entries.

alternatively, we could try deleting the
specific index entity from the cache after
applying an update to the entry - but at
the usage site, we currently do not use
the entity.

#4860