fix rescheduled instances not being removed from the calendar view
we used "isSameEventInstance" to compare an instance from the map
and the original event, which will only be true for the very
first day the event occurs on.

this resulted in the rescheduled instance not being removed
from the original day.

using the generated startTime for the current instance solves
this.

also move the addDays tests to the CalendarUtilsTest since
that's where they live.

fix #5688
fix 3a68f207dd02637603419b2d661d68c46ba6645e