Show a warning in case the user does not have WebAssembly
WebAssembly will be required in a future version of Tutanota, and some
users may have this disabled on their browser. To ensure that users do
not get locked out, the client should warn the user upon logging in that
they will need to enable it or download an app that has it.

Additionally, Apple introduced Lockdown Mode which disables WebAssembly.