[android] fix excessive bundle size when putting app in background
we have been saving the webView state in onSaveInstanceState, which
puts the current history stack into the bundle. Our history entries
can be quite big because they contain the favicon and encoded custom
theme and custom logo.

even on non-whitelabel accounts it could get big by just having a lot
of entries.

we're never restoring this state, so it should be fine to just not save
it and prevent the resulting crashes from bundles > 1MB.

also fix some warnings

fix #5529