Fix applying and displaying system theme
One issue was missing load of theme preference on init so settings did
show the resolved theme id which came from native.

Another issue was not reloading the theme when there was a theme
passed in the URL. Because of that we did not get a correctly
resolved theme on the first start.

fix #5831
fix #5832

Co-authored-by: ivk <ivk@tutao.de>