fix short events not being displayed sometimes in the 6th week row
we changed addDaysForEventInstance to only worry about the month it
was called for and ignore the rest of the event that may extend to the
next month, but did not adapt CalendarViewModel._addOrUpdateEvent
to also add short events to the month where they end.

in cases where we
* receive an entity event update for a short event that starts in the
  current month but ends in the next
* the month view displays a complete week that is part of the next month
  at the bottom (to fill the 6x7 grid)

we would therefore not display that events end.

fix #5896
fix #5839
fix 3a68f207dd02637603419b2d661d68c46ba6645e