fix inconsistent settings links
When the user opens a link that forward to the setting pages the
displayed folder is the Login one no matters the folder that is present
in the path. This happens due the handle of the path be made before the
load of all available folders of the user.

This commit call the path handler after loading all the folders, so it
can successfully handle the requested path.

fix #5515