Do not use the whole buffer for sjcl toBits()
We've been passing the whole underlying buffer of Uint8Array to toBits()
which works for most cases but can break if Uint8Array points only to
the part of the underlying buffer. This is easy to run it with node's
Buffer.from() which allocates a big chunk of data at once so it's
underlying buffer will in many cases bigger than the byteLength.