fix file attachment bubble out of screen
When the email has to attachments on mobile and the two attachments are
placed side by side in the email view, when opens the action bubble for
the right one the bubble overflow to the right clipping part of its
content.

This commit verifies if the bubble will overflow the screen and inverts
the side of it, instead to open to the right, it will open to the left
side of screen.

fix #5919