fix error when search list loads from ID not contained in results
sometimes, the search result display list passes an ID to fetch() that
is not actually contained in the current search result list.
possible causes:
1. that mail was removed from the results lists while the search was
   active
2. the search query changed and the whole result list is invalidated
   before the display list was notified

both cases are not serious errors and can be fixed by re-searching, case
1 is also very unlikely to happen because we only add to the result list
or replace the whole result object.

this commit changes the logic to load from the newest ID in the results
that is older than or equal to the requested load start instead of
requiring an exact match, which should give plausible results at least
in case 1, as well as binding the results object to be fetched from
to the lifetime of the listmodel and checking if it changed before
extending it, which fixes case 2.

fix #5340