fix multiple emails are selected when using long press
When the user is on mobile or more specific on a viewport that uses the
single column layout and visualises an email, when back to mail list and
long press to select another different email, the first one keeps
selected.

This commit changes the behavior of the emails lists and contacts lists,
discarding the previous selected email if in single column layout. On
bigger viewports that uses two or more columns, the behavior stays the
same.

fix #5696