Fix invalid URL being proposed for logging with second factor
appIdToLoginDomain was not taking into account that the Webauthn keys
are always registered for the superdomain and that domain cannot be
used to make a login URL.

Now we special case our superdomains, like in other Webauhn cases.

In addition, we return the URL to login in on instead of returning a
domain and pretending that we can just make a login URL out of it.

We considered adding another parameter to the DomainConfigs but it
seems like it would complicate it even further because (given the
current pattern) we would need to add two URLs to each domain again.

This changes how the URL is displayed to the user in the dialog where
we offer to log in on another domain. We consider this acceptable and
we could convert URL back to the domain if we wish to do so in the
future.

fix #6015