Fix processing calendar event updates when MailDetails is enabled
0cf654d06 changed how we handle missing session key when decrypting
instances which is handled like a programming error now. This broke the
logic for the CalendarModel where we have a temporary solution for not
having a proper key: we are supposed to always pass the key provider
for instances that are encrypted together with Mail but in this case
we don't have a reference to mail yet so we can't do that.

We removed the handling for the missing session key because it
shouldn't result in silent errors on the client and this way we can
actually handle it when we need it like in the case above with calendar
updates.

fix #6021