fix pinch zoom orientation change
The pinch zoom object needs to be recreated
at the right time after the orientation
of the device changes to consider the
right values for the calculation.
for detecting an orientation change of
the device we use a resize listener
on the mail-body / viewport.
The other resize listener on the
zoomable / mail is still needed e.g.
for image loading.

fix #6038