Update Alarm Permission on Android
Starting on API 33, Google stopped granting the SCHEDULE_EXACT_ALARM
automatically when installing the app, making necessary to ask for
user's permission before scheduling alarms.

This commit limits the SCHEDULE_EXACT_ALARM permission to API 32 and
introduces the USE_EXACT_ALARM permission that is granted when
installing the app.

fix #6045