[desktop] Fix making excessive handshakes
Since we switched to built-in fetch() as our implementation we noticed
that a number of users are making excessive TLS handshakes. We've found
that by default undici (that's powering node's built-in fetch()) has
low idle timeout and has no default limit for connections.

We are now explicitly using undici (also more recent version) with
explicitly configured dispatcher to adjust the http client parameters.

We should switch to HTTP2 once it stabilizes in undici.

fix #6053