Implement new navigation for calendar on mobile, change agenda
Add Week/Month preview for previous and next month in Agenda and Day
views.

Add a vertical arrow for showing whether the day selector
is displaying by month or week to the calendar header in the agenda view
on mobile.

Change calendar navigation on mobile, change agenda style

All view types are on the same level now.

View type switcher and today button are in the header.

Month view navigates to agenda.

Mobile agenda only displays events for a single day

If the date from URL is different from today (in desktop) and current
view is Agenda, we move the url to today and show the right information

close #3508
close #6099
fix #5778

Co-authored-by: ivk <ivk@tutao.de>
Co-authored-by: jat <jat@tutao.de>