Show a dialog with failed emails while exporting
When exporting emails there's no indicator of which emails failed to
export.

This commit adds a Dialog informing the user which emails failed to
export and export the emails that had success while downloading,
so users doesn't lose all export progress. The dialog shows the sender
and the subject of the email, allowing the user identify and retry the
failed email.

fix #6131