Fix MAPI does not attach files unless already logged in
When the user is with the desktop app closed and try to send a file
through email the Tuta client doesn't attach the file to the email.

This commit removes the temp folder cleanup at the startup when running
the desktop app, since we might have multiple instances running and
using the temp folder.

fix #5850