Add dark theme based splash screen
When the user have dark mode enabled in the cellphone, the app doesn't
respect the theme while launching the app.

This commit adds a dark theme splash screen and now the app respects the
system active theme while startup.

On Android we've added a new resource under the folders with night
suffix to enable the app determinate which resource use according to
System Theme.

On iOS we've added a new color set (LaunchScreenColorSet) that matches
any and dark theme and also added the dark icon to the Launch Screen
Image Set, with this, we're able to add the dark variation to the
LaunchScreen storyboard

fix #5420