Add Week View to Single Column Layout
Add the Week view to the single column layout and also extract the
expandable date picker to a separate component, allowing further changes
without impacting the current DatePicker.

The time column had its sizes reduced to give more space to the week
view and day view. Now the day selector is able to show the selected day
and today.

Co-authored-by: ivk <ivk@tutao.de>