Highlight week, change today implementation and events for entire week
This commit adds the highlight in day selector for the current week when
expanded and the user is on the week view.

Changes today's button implementation to allow manipulation of it's
content and replacing the inner date dynamically.

Shows the events for the whole week below day selector while on day or
week view and single/two columns layout.

Co-authored-by: ivk <ivk@tutao.de>