[android] Fix crash when push identifier users change
When another user is added or removed from push notifications we try
to restart the connection or stop it if there are none. This was done
on the main thread by accident. We dispatch it to a background thread
now.

Additionally non-successful HTTP responses for SSE connection were not
checked for properly, probably due to the changes in c261a3.

Additionally we were expecting 403 for the authorization failure but
we should have been expecting 401.

fix #6168

Co-authored-by: murilo <mrex@tuta.io>