Trap focus when popup is shown
HTML only allow to trap focus if we use <dialog>, but to do this we've
to render de dialog in a hidden way and call showModal when we need it,
which leads to always have the dialog rendered.

Using inert property in the main-view disable its ability to be
focusable and traps the focus inside the popup.