Fix error when update failed 5 times
When the update fails 5 times, the ElectronUpdater has its listeners
removed, the user receives a notification informing that the software
will try to update later and if the user tries to update again, they
receive an error message.

Since there's no more listeners, and we don't want a massive check for
updates on our servers, the user doesn't receive the error anymore, just
the checkUpdate promise as is resolved as false.

fix #6140