generate invoices on the client
includes sys model v93 for the InvoiceDataService that returns
all data needed to use the PdfInvoiceGenerator to generate a
PDF/A compliant Pdf invoice.

The generator uses the classes in src/api/worker/pdf to write create
the PDF, which contains one high-level imperative PdfDocument that and
defines the document contents and the low level PdfWriter that actually
writes the bytes to a buffer.

tutadb issue 1458