fix text is deleted after closing email template popup
When the user opens the template popup while writing an email and close
the popup without selecting a template, the selected text is replaced
by an empty string.

This commit adds an optional parameter to the TemplatePopup, so when we
close or click outside it without choosing a template, we just use the
optional function to restore the focus and not replace the text by an
empty string.

fix #3263