[ci] silence biome error when committing ignored files
biome has a list of ignored files it will not touch and will be noisy
when explicitly given these files, which is bad for our pre-commit hook.

also doesn't call biome if there's nothing left after filtering out
ignored files.

this is deliberately not sending the error output to /dev/null so we can
see if something really is going wrong.