calendar improvements to larger screen
Changed the calendar layout in larger screens to follow the changes made
in single and two column layout.

Added a day selector in the sidebar, changed the calendar content style
to use cards and toolbar as the Contact View and Mail View.

Added event details to agenda view, so when an event is selected the
event's details is displayed on the right column.

close #6253