optimize ordering of more actions in mail viewer
the "report mail" and "show images" button can both be at the bottom of
the mail action menu.

this moves the "destructive" actions (unsubscribe, show/hide images)
further up so they're not at the bottom of the list of actions
where the report mail action is when it is available.

close #6196