unblock external content in mail editor if there is no external content
the mail editor unconditionally showing the banner to unblock external
content for a new reply/forward if the mailviewer was set to "block"

it also showed the banner if the mail viewer was showing external
content, but there was no explicit rule.

by counting the amount of external content after each sanitization we
can determine if it is actually necessary to block anything.
also fixes the blocking status not being updated to the current value
when restoring a minimized editor.

fix #6239