Fix back navigation in Calendar working incorrectly for day/week view
When the user is on day/week view and press the android back key it
is redirected to month view, even if it never visited it.

This commit removes the handleBackPress from Calendar header, as it
doesn't have to deal with it anymore since we want the same behavior as
we have for month view.

fix #6280