fix wrong undecryptable session key breaking entity loading
we're already sort of ignoring the case where the session key can't be
found for non-nefarious reasons.

this commit adds the case where we encounter a CryptoError during
session key resolution to the ignored errors.

we also catch both CryptoError variants now: the one inheriting from
TutanotaError and the one defined in tutanota-crypto. The latter was
escaping the catch clause because of the instanceof check. This fixes
the websocket not processing new events after encountering this error.

we also harden the indexer against mail that contains _error fields by
skipping them during indexing.