Reconcile CryptoErrors
Introduces a new package in order to avoid tutanota-crypto depending
on the whole tutanota-utils. With this package we can have one canonical
CryptoError.

We add another entry point to tutanota-crypto in order to be able to
import only error classes from any chunk.

We've adjusted moduleResolution for tsconfig to support export maps
for our new entry point.

We've updated @rollup/plugin-typescript to add support for export maps.

We have reorganized utils to adjust the chunking. Error utils are now
a part of the main chunk.

We have moved ServiceWorkerClient into the main chunk because we need
to handle the errors and also because it is anyway invoked only after
the other initialization is done.

close #6342