Add keyboard control to `InfoIcon`
ARIA does not have a standard for tooltip keyboard controls, so I chose
to show the tooltip on focus. This is implied to be the standard by the
guidelines via the page for the tooltip role on WAI-ARIA 1.3. It is also
implied by a tooltip not being a pop-up under the aria-haspopup
specification on WAI-ARIA 1.3.