Fix mails disappear from search view when moved
When the user search for emails and another client do some update that
affects the current results, the affected emails where gone from the
list and the search needs to be done again.

This commit handles the MOVE operation by checking if it really is a
MOVE operation (the updates contain a CREATE and DELETE for same id) and
then handle to update the current states to match the new listId.

fix #1732