Fix creating a template list causes SessionKeyNotFoundError
Seems that the problem have already been fixed, but this commit
implements the changes suggested in the issue.

Also, fixes a bug that the contactAdmin_msg was displayed even when the
user was the global admin or user was allowed to create a template. This
wasn't preventing real allowed users from creating templates.

close #6205