fix search view missing some updates to the search URL
the list that is recreated when the URL changes also calls
updateSearchUrl which then overrides later state changes from the search
bar. re-ordering some operations that happen when clearing the search
bar or routing from quicksearch fixes that.

also cancel the current search when leaving the search view or changing
the search bar content.