Fix being unable to disable mail search index in apps
There are multiple bugs that all caused this.

1. There was a loop where Indexer.disableMailIndexing() would call
init() which would re-initialize mail indexing again. We moved enabling
mail indexing out of init() as it was mixing up responsibilities.  We
could make init() return whether it is a new database but instead  we
rely on isNewOfflineDb which should work the same way.
2. OfflineStorage.init() was not checking for the last update correctly
which made all the places that check for isNewOfflineDb work
incorrectly.
3. MailSettingsViewer was still passing a reason string in place of a
user id which made it impossible to disable search index. We've removed
the parameter as it is redundant and inconsistent.

fix #6412