Enforce MAC checking where possible
In general we try to enforce MACs whenever we use AES-256, but in some cases we cannot do that because of legacy data.
We had already been encrypting credentials, index and some configurations (all locally) with 256 bit keys but without MACs, so we will need a migration for that later.

Another important change is always using MACs when encrypting RSA keys. There is no reason not to do this, even though we plan to get rid of those soon.

Additionally, this introduces some general code improvements, such as:
- type safety on Android
- rename variables to make clear which keys they refer to
- make return types explicit

tutadb#1669

Co-authored-by: @vitoreiji