Sync contacts after receiving EntityEventUpdates
This commit implements the sync of contacts after receiving an
EntityUpdate, allowing the users to have their contact book updated
not only while startup, but also during runtime.

The operation of delete was also implemented, now the app is able to
delete contacts from the contact book at startup or during runtime.