[ios] Add `swift-format` as formatter
I decided on using Apples Swift formatter, `swift-format` as it is
planned to be included in future versions of the Swift toolchain & its
style appears to be the most popular online from a quick glance.

The SwiftLint configuration was also changed as it includes some
formatting rules by default that overlap with `swift-format`s rules.