update sys model, finish MailDetails migration
* stop setting the offline db version to 0 for old databases
  all offline dbs in use now have this field set to 1 already or are
  outdated and there is no migration that we need to run anymore.
* use raw key for desktop offline db instead of passphrase. this will
  recreate the offline DB for desktop clients.
* update minimum compatible model versions for the offline migrator
* remove old offline migrations. the remaining ones are the ones added
  with the lastNonOutdatedClientVersion (3.118.12 since ~120 days) or
  later, so no client that uses an offline db from before that would be
  able to log in and an offline db left by such a client will not be
  migrated due to being outdated.

In order to finish the migration to MailDetails we need to delete all
mails (with and without mailDetails) and mailBodies.
This is needed as we do no longer store the mailBodies directly on the
mail, and we currently do not support updating specific ranges of mails
that exist in the offline storage.
Therefore, deleting only those mails that have no mailDetails set is not
possible, as available offline ranges for mails might be broken
afterward.

includes sys model 94 for supportInfo on customerInfo.

@co-authored-by: jhm@tutao.de