Migrate to modern KDF on login
The current modern KDF is Argon2id, but the migration can be reused for future KDFs.

This includes
- when creating a new session
- when resuming a session

Also migrate locally stored credentials.

We reload user before verifying the password because we don't want to use an old, cached verifier, otherwise we can get logged out unnecessarily.

In addition, we put in place measures to hold back the migration until all clients can handle it.

tutadb#1655