[android] Add AccountManager on Android
This commit implements the AccountManager to Android, so our contacts
doesn't get deleted by the ContactProvider. Adding this AccountManager
also removes the warning message about synchronization that was
displayed at Google's Contacts app. As an extra, now Tuta contacts
are displayed as belonging to Tuta app, with icon and the related email
address.