Fix contacts not removed when deleting credentials
When deleting an account, if it had contact sync enabled, the contacts
still in the phone book because we were deleting by userId and since we
recently changed to use user's login email address, the native side
couldn't find the account and the contacts were not removed. This commit
fixes it and use User's login email address to remove contacts when
removing an account.