Fix iOS crashing when enabling contact synchronization
When the user tries to enable contact synchronization the app crashes
without error message. The reason was a missing Phone type that was added
with this commit. Also handles errors of missing contact permissions and
displays a message about it, the same way that is done with Android now.

fix #6500