[ios] Do not warn on convenience_type
The intent is obvious: we don't want to be able to instantiate objects
for classes that only have static members.

However, the suggested way to fix this is to declare these classes as
caseless enums, which miscommunicates the intent of the class when it is
because the class purely has static methods (which makes no sense for an
enum).

This will cause more confusion than the (highly unlikely) event that
someone tries to instantiate a FileUtils for some reason, which will be
caught in a code review.

Linters should point out existing cases of user error, not future cases
of user error, which may not even happen.