Add view file handling from native side to app's web part
Now the app is capable to handle native side calls asking for importing
files, mainly used when a user view a supported file and want that Tuta
app handles it.

Currently, the only supported format is .vcf and only on Android, since
iOS doesn't support setting our app as a handler for vcf files.