Align all day events in the calendar on desktop
The ideal solution in my eyes is to rewrite the calendar header so that
the header and body for the day are in one column flex. Using flex
is better than constantly calculating the sizes of elements ourselves
both in code simplicity and performance.

Until then, this quick workaround seems to do the trick.