Add the illustrations to the onboarding wizard
This needs some optimization as the illustrations cause the CSS
animations to lag. `Icons.ts` is not the best place to have the
illustrations so we should create a new file. We could also look at
sanitizing the SVGs once a way to keep the CSS classes in the SVG is
found.