Add an options to select the behavior after moving an email (#6671)
* Add an options to select the behavior after moving an email

This commit adds a new option inside the settings so users can select
the desired behavior after moving or deleting an email while it's being
displayed.

Co-authored-by: wec43 <wec@tutao.de>

Close #6583

* Moves the DeviceConfig out of ListModel and makes the behavior optional

* Makes autoSelectBehavior mandatory in ListModel

* Fix ListModel selection behavior

When deleting the last item in the list and the behavior was to pick the
next older one, the list was getting the first item in the list. This
commit makes the list pick the item right before the last one.