Fix an issue with system notification emails
15713f1 changed to how we detect and display system notifications. Since
tuta-crypt we send all system notifications from the system group
address. It was attempting then to replace the sender with reply-to
address.

This cannot work because system notifications are sent with MailDetails
now, and we do not know authStatus nor replyTos until we load
MailDetails.

This commit simplifies the handling by not attempting to replace the
sender but by hiding it for all system notification mails. Now there
will be more cases when they are hidden then before but it should have
no effect since sender of system notification emails is irrelevant and
replyTos will be used to write a reply.

fix #6708