[ios] Fix contact hashing
The original contact hashing used to check contact changes was based on
Swift's implementation of Hashable that unfortunately is granted to be
the same just during the program execution, leading to different hashes
every app startup.

This commit replaces the use of Hashble by an implementation of
MurmurHash, leading to stable hashes across multiples app startup.

Co-authored-by: ivk <ivk@tutao.de>

fix #6776