Fix calendar issues (#6701)
* Use short names in week view if two months

* Make current date more contrasted

Also move over the styling to classes rather than having three copies of
it!

* Improve how well the mini calendar fits

Remove the padding around the calendar, and increase the minimum size of
the first column from 240px to 270px.

* Always reserve width for the scrollbar for left column in calendar

We use scrollbar-gutter-stable-or-fallback to do this, ensuring that
the width will always be reserved. For the New event button, it will
also align it.

* Move week indicator to the header row for calendar

Remove week from the CalendarNavConfiguration as it is not needed here.

* Make width of the highlighted week not relative

* Fix the sidebar disappearing too early

Fix week titles

Co-authored-by: wrd <wrd@tutao.de>

---------

Co-authored-by: paw <paw-hub@users.noreply.github.com>
Co-authored-by: jat <jat@tutao.de>
Co-authored-by: wrd <wrd@tutao.de>