Fix the `DropDown`'s navigation and it's filter including cosmetic periods (#6812)
* Combine the `DropDownButton` and the 'Add Folder' button

* Fix `DropDown`'s keyboard navigation

This focuses the filter on open and makes using the arrow buttons to
focus an item smoother.

* Fix matching indented folder names in the drop-down

This excludes the periods from the filter search except when they are
in the name of the folder itself.

* Fix screen readers announcing the move folder dropdown incorrectly

Removing the `title` attribute fixes the name of the folder being
announced twice. This is fine as `title` is meant for pop-ups according
to MDN while it was being used to add the label again.

* Stop the filter input from being focused when the dropdown opens