Fix dropdown filtering by wrong field (#6852)
Label is a human-readable accessibility description, and it might
include parts that we don't want to filter by.

We opted into preferring text to label because that's what users are
more likely to see. It is not perfect as in some cases (e.g. folder
names) we display something that we don't want to filter by (folder
indentation level with dots) but it is less surprising. A proper fix
would be to introduce another value/function for filtering.

#5353

Co-authored-by: wrd <wrd@tutao.de>