Fix the list's rows not being highlighted on tab
There is still a possible issue where the tab order and the selection
via the arrow keys do not align. If you tab to the third row then
press down for instance, the forth row will not be selected.

I am not sure whether to fix this as the arrow keys and tab do slightly
different things. The arrow keys focus and open a row while the tab just
focuses a row.