Show pq padlock symbol for sent mail, tutadb 1788 (#6699)
* Show pq padlock symbol for sent mail, tutadb 1788

To show the pq padlock symbol for the sender we introduced another enum value for the encryptionAuthStatus on Mail instances which is only set for the sent mail. It is only set in case the mail is confidential and all recipients support the TutaCrypt protocol.

* clean up outdated invitations from offlince cache
see the fix for tutadb#1800

---------

Co-authored-by: bedhub <bedhub@users.noreply.github.com>
Co-authored-by: vaf <vaf@tutao.de>