[ios] Block multiple AppStore subscriptions
AppStore allows only one subscription per AppleId, so this commit checks
and block if a user tries to upgrade another Tuta account with the same
AppleId. Also ensure that the user is the current owner of the
subscription, preventing a user from paying twice for the same Tuta
account w/ different AppleIds.