Update user when the one we have is out-of-date
When loading a key at a certain version, if we notice the user has a key for that group with a lower version, this must mean that the memberships are out-of-date, because they should always have the latest key versions.

When this happens, we force downloading the user and update it on the facade, then we try getting the key again.

tutadb#1770