Fix iOS recognizing other files as vCard when sharing
When sharing files on iOS the file can be ignored if its extension isn't
included inside a switch case that determinate how it should be loaded.
This commit sets the fallback behavior to try loading the file instead
of returning null directly.

The problem of iOS suggesting contact import for some files was related
with the issue described above. When it fails to load the attachment
the web part performs an every check on an empty array looking files
with vCard mime-type and the default return value for empty arrays is
true, so the app was considering an empty array as an array of vCard
files.

fix #6966