Implement address-field canvas rendering
Allow the PdfInvoiceGenerator to generate an address field containing
any character by utilizing the OffscreenCanvas API.

This allows going beyond the limited 256 glyphs provided without using
PDF CID fonts. The canvas produces an image which we embed into the PDF.
This is not particularly accessible, but sufficient before a better
solution can be found.

If the canvas rendering fails or if there's no need to render with the
canvas, the implementation will fall back to normal text rendering.

If a canvas is rendered, ghost text will be printed below it to give
partial accessibility

closes #7075