Make key resolution more robust
tutadb#1826

There are multiple error cases where entity rest cache and key
cache maybe out of sync or still missing some updates. this is
hard to prevent when processing entity updates. therefore, we try
our best to reload the respective entities from the server and
refresh the key cache.

To reduce complexity we accept that there may still be some corner
cases that fail in unspecified ways such as when we are the member
of a shred group that was rotated and have not yet processed the
key update or when we are the admin and are slow in getting key
rotation from the user.