[Calendar] Removes default reminders from shared calendars
This commit removes the recently added feature that allows users to set multiple default reminders from not owned shared calendars.
These reminders are unique for each user's calendar and should not be available for shared calendars the user don't own.

CalendarInfo.userIsOwner was introduced, so we could keep using the CalendarInfo.shared field to preserve the current behavior(v235.240718.0) when dealing with selecting calendars.
This commit also rewrites some of the logic related to CalendarInfo.shared, which affects calendar sharing, calendar list(on the sidebar) and event invites,especially the capabilities a user has when editing an event.