fix moving to maildetails breaking old search indexes
previously, we had some associations directly on Mail with different
ids than in the MailDetails model, which means that old
index entries will not be able to be used with the new association ids.
MailWrapper was masking this problem by returning the values from the
new types in places where Mail.toRecipients et al were asked for.
With the removal of MailWrapper we need to hardcode these old ids.

this workaround does not have to be maintained in case the mail indexes
have been dropped after 2024-07