Fix contact selection issues
This is another spin on solving #7256.

This reverts 5be83674 and 26dfa0f5 as they introduced some
issues with initialization, namely `initOnce` could run
multiple times.

In this commit we try to acknowledge that existing URL pattern is not
sufficient to distinguish between the cases as we might have selection
without focus.

Instead, we introduce an explicit argument to focus the item details.

We considered synchronizing the column focus state using this argument
in all cases but this would entail a larger refactoring.