Fix the credit card verification
This fixes the currency symbol being included and the decimal separator
in the call to the Braintree card verification causing it to fail.

It is not possible to have `getSubscriptionPriceWithCurrency` return
without currency formatting because the return of `getPlanPrices` cannot
be changed as `StoreKit` can only return the price in the
local currency.

Closes #7370.

Co-authored-by: wec <wec@tutao.de>