Fix group members being kicked out when rotating keys
When encrypting asymmetrically for recipients from another customer, we
pass in a list of recipients for whom we already could not find a public
key, and if that list is not empty, no encryption will take place
because we just want to collect the recipients we cannot find. In the
case of rotation, though, we need the encryption to take place, so we
always pass in an empty list, and we handle the recipients without
public keys separately with a different list.

tutadb#1839