Adds new App colors
This commit adds two new colors to Mail and Calendar app and removes the
blue theme, replacing it by the new Blue theme using Calendar Colors.

This commit also adapts how the themes are displayed, Light and Dark can
be Light Blue/Red and Dark Blue/Red this will depend on which app the
user is running.